feat: add tournament simulation service and adaptive match outcomes (#2)
- Added SimulationService to simulate group stage and knockout rounds
- Supported winner-only, fixed-result, and adaptive probability simulations across tournament phases
- Updated AttackRating and DefenseRating calculators to return AttackRatingDTO and DefenseRatingDTO from Calculate methods
- Added PickRandomScore and PickRandomOutcome methods in MatchProbabilityService
- Added MatchResultDto, SimpleMatchResultDto, AdaptativeSimulationMatchDto, SimpleSimulationMatchDto, and AdaptativeMatchResultDto for simulation flows
- Changed AccumulativeCount from double to int in NationalTeam
- Added migrations for simulation and rating model updates

+++ WorldCupSimulatorApp/WCS.Application/DTO/RatingsDTO/DefenseRatingDTO.cs
@@ -4,6 +4,6 @@
     {
         public double DefenseRating { get; set; }
         public double AccumulatedPenalties { get; set; }
-        public double AccumulatedCount { get; set; }
+        public int AccumulatedCount { get; set; }
     }
 }
